finish new input code
