--- react/InputNew.js
@@ -1,3 +1,0 @@
-import InputNew from './components/Form/InputNew/index'
-
-export default InputNew
