--- react/components/Form/CurrencyInput/style.css
@@ -1,8 +1,0 @@
-.input:focus {
-  outline: 0;
-  /* border-color: #4B71FC; */
-}
-
-.input::placeholder {
-  color: #8BA3AA;
-}
