+++ react/components/Form/Input/style.css
@@ -1,8 +1,11 @@
 .input:focus {
   outline: 0;
-  /* border-color: #4B71FC; */
 }
 
 .input::placeholder {
   color: #8BA3AA;
+}
+
+.inputIconBefore {
+  margin-top: -1px;
 }
