--- react/components/Form/InputNew/style.css
@@ -1,11 +1,0 @@
-.input:focus {
-  outline: 0;
-}
-
-.input::placeholder {
-  color: #8BA3AA;
-}
-
-.inputIconBefore {
-  margin-top: -1px;
-}
